Move types/tx to types
