+++ x/feegrant/internal/types/codec.go
@@ -10,6 +10,8 @@ func RegisterCodec(cdc *codec.Codec) {
 	cdc.RegisterInterface((*exported.FeeAllowance)(nil), nil)
 	cdc.RegisterConcrete(&BasicFeeAllowance{}, "feegrant/BasicFeeAllowance", nil)
 	cdc.RegisterConcrete(&PeriodicFeeAllowance{}, "feegrant/PeriodicFeeAllowance", nil)
+
+	cdc.RegisterConcrete(FeeGrantTx{}, "cosmos-sdk/FeeGrantTx", nil)
 }
 
 // ModuleCdc generic sealed codec to be used throughout module
