+++ x/feegrant/internal/types/test_common.go
@@ -1,5 +1,6 @@
 // noalias
-package tx
+
+package types
 
 import (
 	"github.com/tendermint/tendermint/crypto"
