--- x/feegrant/internal/types/tx/codec.go
@@ -1,8 +1,0 @@
-package tx
-
-import "github.com/cosmos/cosmos-sdk/codec"
-
-// RegisterCodec registers concrete types on the codec
-func RegisterCodec(cdc *codec.Codec) {
-	cdc.RegisterConcrete(FeeGrantTx{}, "cosmos-sdk/FeeGrantTx", nil)
-}
